# flutter/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Travel Memory Vault API",
    "description": "Travel Memory Vault - Personal travel journal with AI-powered search",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/trips": {
      "post": {
        "tags": [
          "Trips"
        ],
        "summary": "Create Trip",
        "description": "Create a new trip.\n\n    **Authentication:** Required\n\n    **Permissions:** Any authenticated user\n\n    **Request Body:**\n    - title: Trip title (required, 1-255 chars)\n    - description: Optional trip description\n    - start_date: Optional trip start date (YYYY-MM-DD)\n    - end_date: Optional trip end date (YYYY-MM-DD)\n    - cover_photo_url: Optional cover photo URL\n    - visibility: private | unlisted | public (default: private)\n\n    **Returns:**\n    Created trip with generated ID\n\n    **Response Example:**\n```json\n    {\n        \"id\": \"123e4567-e89b-12d3-a456-426614174000\",\n        \"user_id\": \"987e6543-e21b-12d3-a456-426614174000\",\n        \"title\": \"Summer Europe Trip\",\n        \"description\": \"Backpacking across Europe\",\n        \"start_date\": \"2025-06-01\",\n        \"end_date\": \"2025-06-30\",\n        \"visibility\": \"private\",\n        \"views_count\": 0,\n        \"saves_count\": 0,\n        \"created_at\": \"2025-01-25T10:30:00Z\",\n        \"updated_at\": \"2025-01-25T10:30:00Z\"\n    }\n```\n\n    **Errors:**\n    - 400: Invalid date range (end_date < start_date) or invalid visibility\n    - 401: Not authenticated\n    - 403: Free tier limit reached (6 trips max)\n\n    **Business Logic:**\n    - user_id is automatically set from authenticated user\n    - Free tier users: max 6 trips\n    - Premium users: unlimited trips\n    - Default visibility is \"private\"\n    - views_count and saves_count initialized to 0",
        "operationId": "create_trip_api_v1_trips_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Bearer token from Supabase Auth",
              "title": "Authorization"
            },
            "description": "Bearer token from Supabase Auth"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TripCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Trips"
        ],
        "summary": "List Trips",
        "description": "List trips with pagination.\n\n    **Authentication:** Required\n\n    **Permissions:** Any authenticated user\n\n    **Query Parameters:**\n    - page: Page number (default: 1, min: 1)\n    - page_size: Items per page (default: 20, max: 100)\n    - visibility: Optional filter by visibility (private|unlisted|public)\n    - public_only: When true, returns global public trips for social feed\n\n    **Returns:**\n    Paginated list of trips with metadata\n\n    **Response Example:**\n```json\n    {\n        \"trips\": [\n            {\n                \"id\": \"123e4567-e89b-12d3-a456-426614174000\",\n                \"title\": \"Summer Europe Trip\",\n                ...\n            },\n            {\n                \"id\": \"987e6543-e21b-12d3-a456-426614174000\",\n                \"title\": \"Winter Japan Trip\",\n                ...\n            }\n        ],\n        \"total\": 15,\n        \"page\": 1,\n        \"page_size\": 20,\n        \"total_pages\": 1\n    }\n```\n\n    **Errors:**\n    - 401: Not authenticated\n\n    **Business Logic:**\n    - Default (`public_only=false`): returns only trips owned by current user\n    - Social feed (`public_only=true`): returns only public trips from all users\n    - Results ordered by created_at DESC (newest first)\n    - Page size automatically capped at 100\n    - Empty list if user has no trips",
        "operationId": "list_trips_api_v1_trips_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Page number (1-indexed)",
              "default": 1,
              "title": "Page"
            },
            "description": "Page number (1-indexed)"
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Items per page (max 100)",
              "default": 20,
              "title": "Page Size"
            },
            "description": "Items per page (max 100)"
          },
          {
            "name": "visibility",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by visibility (private|unlisted|public)",
              "title": "Visibility"
            },
            "description": "Filter by visibility (private|unlisted|public)"
          },
          {
            "name": "public_only",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "When true, return only public trips across all users",
              "default": false,
              "title": "Public Only"
            },
            "description": "When true, return only public trips across all users"
          },
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Bearer token from Supabase Auth",
              "title": "Authorization"
            },
            "description": "Bearer token from Supabase Auth"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/trips/{trip_id}": {
      "get": {
        "tags": [
          "Trips"
        ],
        "summary": "Get Trip",
        "description": "Get trip detail by ID.\n\n    **Authentication:** Required\n\n    **Permissions:**\n    - Owner can always view their trip\n    - Non-owners can view if trip is public or unlisted\n    - Private trips only visible to owner\n\n    **Path Parameters:**\n    - trip_id: Trip UUID\n\n    **Returns:**\n    Trip details\n\n    **Response Example:**\n```json\n    {\n        \"id\": \"123e4567-e89b-12d3-a456-426614174000\",\n        \"user_id\": \"987e6543-e21b-12d3-a456-426614174000\",\n        \"title\": \"Summer Europe Trip\",\n        \"description\": \"Backpacking across Europe\",\n        \"start_date\": \"2025-06-01\",\n        \"end_date\": \"2025-06-30\",\n        \"visibility\": \"public\",\n        \"views_count\": 142,\n        \"saves_count\": 23,\n        \"created_at\": \"2025-01-25T10:30:00Z\",\n        \"updated_at\": \"2025-01-25T10:30:00Z\"\n    }\n```\n\n    **Errors:**\n    - 401: Not authenticated\n    - 403: Trip is private and user is not owner\n    - 404: Trip not found\n\n    **Business Logic:**\n    - Increments views_count if viewer is NOT the owner\n    - Access control based on visibility:\n      - private: Only owner can view\n      - unlisted: Owner + anyone with link can view\n      - public: Anyone can view",
        "operationId": "get_trip_api_v1_trips__trip_id__get",
        "parameters": [
          {
            "name": "trip_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Trip Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Bearer token from Supabase Auth",
              "title": "Authorization"
            },
            "description": "Bearer token from Supabase Auth"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Trips"
        ],
        "summary": "Update Trip",
        "description": "Update an existing trip.\n\n    **Authentication:** Required\n\n    **Permissions:** Only trip owner can update\n\n    **Path Parameters:**\n    - trip_id: Trip UUID\n\n    **Request Body:**\n    All fields are optional (partial update):\n    - title: New trip title (1-255 chars)\n    - description: New description\n    - start_date: New start date\n    - end_date: New end date\n    - cover_photo_url: New cover photo URL\n    - visibility: New visibility (private|unlisted|public)\n\n    **Returns:**\n    Updated trip\n\n    **Response Example:**\n```json\n    {\n        \"id\": \"123e4567-e89b-12d3-a456-426614174000\",\n        \"title\": \"Updated Trip Title\",\n        \"description\": \"Updated description\",\n        ...\n    }\n```\n\n    **Errors:**\n    - 400: Invalid date range (end_date < start_date)\n    - 401: Not authenticated\n    - 403: User does not own trip\n    - 404: Trip not found\n\n    **Business Logic:**\n    - Only updates provided fields (partial update)\n    - Ownership check prevents unauthorized updates\n    - Date validation ensures end_date >= start_date\n    - Cannot update user_id or engagement metrics (views, saves)\n    - updated_at timestamp automatically updated",
        "operationId": "update_trip_api_v1_trips__trip_id__patch",
        "parameters": [
          {
            "name": "trip_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Trip Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Bearer token from Supabase Auth",
              "title": "Authorization"
            },
            "description": "Bearer token from Supabase Auth"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TripUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Trips"
        ],
        "summary": "Delete Trip",
        "description": "Delete a trip.\n\n**Authentication:** Required\n\n**Permissions:** Only trip owner can delete\n\n**Path Parameters:**\n- trip_id: Trip UUID\n\n**Returns:**\n204 No Content on success\n\n**Errors:**\n- 401: Not authenticated\n- 403: User does not own trip\n- 404: Trip not found\n\n**Business Logic:**\n- Ownership check prevents unauthorized deletion\n- Cascades to related records:\n  - All trip_places deleted (via ON DELETE CASCADE)\n  - All trip_routes deleted (via ON DELETE CASCADE)\n- Permanent deletion (no soft delete)\n\n**Warning:**\n- This action is irreversible\n- All associated places and routes will be deleted",
        "operationId": "delete_trip_api_v1_trips__trip_id__delete",
        "parameters": [
          {
            "name": "trip_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Trip Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Bearer token from Supabase Auth",
              "title": "Authorization"
            },
            "description": "Bearer token from Supabase Auth"
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/trips/{trip_id}/bounds": {
      "get": {
        "tags": [
          "Trips"
        ],
        "summary": "Get Trip Bounds",
        "description": "Get geographic bounding box for a trip based on its places.\n\n    **Authentication:** Required\n\n    **Permissions:**\n    - Trip owner can always view\n    - Non-owner can view if trip is public/unlisted\n\n    **Path Parameters:**\n    - trip_id: Trip UUID\n\n    **Returns:**\n    Bounding box with min/max coordinates and center point, or null if no places\n\n    **Response Example:**\n```json\n    {\n        \"min_lat\": 48.8530,\n        \"min_lng\": 2.2945,\n        \"max_lat\": 48.8738,\n        \"max_lng\": 2.3499,\n        \"center_lat\": 48.8634,\n        \"center_lng\": 2.3222\n    }\n```\n\n    **Null Response (no places):**\n```json\n    null\n```\n\n    **Errors:**\n    - 401: Not authenticated\n    - 403: Trip is private and user is not owner\n    - 404: Trip not found\n\n    **Business Logic:**\n    - Calculates bounding box from all places in trip\n    - Returns null if trip has no places\n    - Access control based on trip visibility\n    - Useful for auto-centering map on trip\n    - Center is simple average of bounds (not geographic centroid)\n\n    **Use Cases:**\n    - Auto-fit map to show all places in trip\n    - Calculate zoom level for trip map view\n    - Determine trip geographic span",
        "operationId": "get_trip_bounds_api_v1_trips__trip_id__bounds_get",
        "parameters": [
          {
            "name": "trip_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Trip Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Bearer token from Supabase Auth",
              "title": "Authorization"
            },
            "description": "Bearer token from Supabase Auth"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TripCreate": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "start_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Date"
          },
          "end_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Date"
          },
          "cover_photo_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cover Photo Url"
          },
          "visibility": {
            "type": "string",
            "title": "Visibility",
            "default": "private"
          },
          "activity_focus": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Activity Focus"
          },
          "travel_style": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Travel Style"
          },
          "budget_category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Budget Category"
          }
        },
        "type": "object",
        "required": [
          "title"
        ],
        "title": "TripCreate",
        "description": "Trip creation schema.\n\nInherits all fields from TripBase.\nUsed for POST /trips endpoint.\n\nBusiness Logic:\n    - user_id is automatically set from authenticated user\n    - Free tier users: max 6 trips (enforced in service layer)\n    - Premium users: unlimited trips"
      },
...
      "TripResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "start_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Date"
          },
          "end_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Date"
          },
          "cover_photo_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cover Photo Url"
          },
          "visibility": {
            "type": "string",
            "title": "Visibility"
          },
          "views_count": {
            "type": "integer",
            "title": "Views Count"
          },
          "saves_count": {
            "type": "integer",
            "title": "Saves Count"
          },
          "place_count": {
            "type": "integer",
            "title": "Place Count",
            "default": 0
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "title",
          "visibility",
          "views_count",
          "saves_count",
          "created_at",
          "updated_at"
        ],
        "title": "TripResponse",
        "description": "Trip response schema.\n\nAttributes:\n    id: Trip UUID\n    user_id: Owner user ID\n    title: Trip title\n    description: Trip description\n    start_date: Trip start date\n    end_date: Trip end date\n    cover_photo_url: Cover photo URL\n    visibility: Visibility setting\n    views_count: Number of views\n    saves_count: Number of saves\n    created_at: Creation timestamp\n    updated_at: Last update timestamp\n\nConfig:\n    from_attributes: Enable ORM mode for SQLAlchemy models"
      },
...
    }
  }
}